scheduler/requirements.txt
requests
apscheduler
python-dotenv